+++ ethereum/test/beefy/gas.ts
@@ -1,13 +1,13 @@
-import { ethers, expect, loadFixture, mine } from "./setup"
-import { baseFixture } from "./fixtures/beefy"
+import { ethers, expect, loadFixture, mine } from "../setup"
+import { baseFixture } from "./fixtures"
 import {
     createRandomPositions,
     createValidatorFixture,
     createInitialValidatorProofs,
     createFinalValidatorProofs,
-} from "./helpers"
+} from "../helpers"
 
-import fixtureData from "./fixtures/data/beefy-relay-basic.json"
+import fixtureData from "../data/beefy-relay-basic.json"
 
 let SUBMIT_FINAL_2 =
     "submitFinal(uint256,(uint32,uint64,(bytes32,bytes,bytes)),(bytes[],uint256[],address[],bytes32[][]),(uint8,uint32,bytes32,uint64,uint32,bytes32,bytes32),(bytes32[],uint64))"
