+++ ethereum/test/channels/basic_outbound_channel.test.ts
@@ -1,5 +1,5 @@
-import { ethers, expect, loadFixture } from "./setup"
-import { BasicOutboundChannel__factory } from "../src"
+import { ethers, expect, loadFixture } from "../setup"
+import { BasicOutboundChannel__factory } from "@src"
 
 describe("BasicOutboundChannel", function () {
     let testPayload = ethers.utils.formatBytes32String("arbitrary-payload")
