+++ ethereum/test/libs/data/mmr-fixture-data-7-leaves.json
